Add Axis Font Options & Fix Axis/Legend Tests
Add Methods to Control Font on chart axis
Also noticed when adding tests for new methods that the Legend and Axis
Tests were misnamed.
